修改vps清除账号报错

+++ follow-order-mascontrol/src/main/java/net/maku/mascontrol/controller/FollowVpsController.java
@@ -289,7 +289,10 @@ public class FollowVpsController {
         List<FollowTraderEntity> list = followTraderService.list(new LambdaQueryWrapper<FollowTraderEntity>().eq(FollowTraderEntity::getServerId, vpsId));
         List<Long> idList = list.stream().map(FollowTraderEntity::getId).toList();
         //删除跟单关系
-        followTraderSubscribeService.remove(new LambdaQueryWrapper<FollowTraderSubscribeEntity>().in(FollowTraderSubscribeEntity::getMasterId, idList).or().in(FollowTraderSubscribeEntity::getSlaveId, idList));
+        if(ObjectUtil.isNotEmpty(idList)) {
+            followTraderSubscribeService.remove(new LambdaQueryWrapper<FollowTraderSubscribeEntity>().in(FollowTraderSubscribeEntity::getMasterId, idList).or().in(FollowTraderSubscribeEntity::getSlaveId, idList));
+        }
+
         //删除账号
         followTraderService.remove(new LambdaQueryWrapper<FollowTraderEntity>().eq(FollowTraderEntity::getServerId, vpsId));
         return Result.ok();
